plugins/utilization_tool/pyproject.toml
[build-system]
requires = ["setuptools==80.9.0"]
build-backend = "setuptools.build_meta"

[project]
name = "auditor_utilization_plugin"
version = "0.10.1"
requires-python = ">=3.9"
dependencies = [
	     "python-auditor==0.10.1",
	     "cryptography==46.0.3",
	     "pyyaml==6.0.3",
	     "pydantic==2.12.3",
       "dotenv==0.9.9",
       "numpy==2.0.2",
       "pandas==2.2.3",

]
description = "AUDITOR plugin to create cluster utilization plugin"
readme = "README.md"

[project.optional-dependencies]
style = [
      "black==25.9.0",
      "ruff==0.14.2",
]
tests = [
      "pytest==8.4.2",
      "pytest-cov==7.0.0",
      "mypy==1.18.2",
      "types-pyyaml==6.0.12.20250915",
      "pytest-asyncio==0.24.0",
      "pandas-stubs==2.2.2.240807",
      "types-python-dateutil==2.9.0.20260124",

]
build = [
      "build==1.3.0",
      "setuptools==80.9.0",
]

[project.scripts]
auditor-utilization-plugin = "auditor_utilization_plugin.main:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.coverage.lcov]
output = "lcov.info"

[tool.black]
include = '\.pyi?$'

[tool.pytest.ini_options]
addopts = "-s -v --cov --cov-report lcov"
testpaths = ["tests"]
filterwarnings = ["ignore::DeprecationWarning"]

[[tool.mypy.overrides]]
module = ["pyauditor"]
ignore_missing_imports = true

[tool.ruff.lint]
extend-select = ["I"]

